+++ src/message.cc
@@ -31,7 +31,7 @@ bool operator<(const Message& lhs, const Message& rhs) {
 
 std::ostream& operator<<(std::ostream& o, const Message& m) {
   o << "{id: " << m.round << ", order: " << OrderString(m.order) << ", ids: <";
-  for (int i = 0; i < m.ids.size(); ++i) {
+  for (size_t i = 0; i < m.ids.size(); ++i) {
     if (i > 0) o << ' ';
     o << m.ids[i];
   }
